Add annual rotation of pisg stats
Introduce an annual job that will rotate the stats at the end of each year.

The rotation runs after the last stats generation of the previous year and
moves the stats HTML and dependent image files to a subdirectory named after
the previous year.

A cron job is maintained to run the rotate-pompone-pisg.yml playbook.

Logs are archived to the bot-dir/archives directory.  These archives are
included in scheduled backups.